interactor set state tests moved to multiversx-sc-snippets
