+++ framework/snippets/.gitignore
@@ -1,0 +1,2 @@
+# Test output
+set_state.json
